+++ src/steps/cloud-asset/index.ts
@@ -1,4 +1,3 @@
-import { uniq } from 'lodash';
 import {
   createDirectRelationship,
   createMappedRelationship,
@@ -37,12 +36,7 @@ import {
   createIamBindingEntity,
 } from './converters';
 import get from 'lodash.get';
-import {
-  GOOGLE_RESOURCE_TO_J1_TYPE_MAP,
-  J1_TYPE_TO_KEY_GENERATOR_MAP,
-  NONE,
-  traverseCloudResourcesMap,
-} from '../../constants';
+import { getTypeAndKeyFromResourceIdentifier } from '../../utils/iamBindings/getTypeAndKeyFromResourceIdentifier';
 
 export async function fetchIamBindings(
   context: IntegrationStepContext,
@@ -244,100 +238,42 @@ export async function createPrincipalRelationships(
   );
 }
 
-function getTargetKey(targetResourceType: string, resource: string) {
-  const keyGenFunction = J1_TYPE_TO_KEY_GENERATOR_MAP[targetResourceType];
-  if (!keyGenFunction) {
-    console.log(
-      { resource, targetResourceType },
-      'unable to find key generation function for J1 type.',
-    );
-    return;
-  }
-  return keyGenFunction(resource);
-}
-
 export async function createMappedBindingAnyResourceRelationships(
   context: IntegrationStepContext,
 ): Promise<void> {
-  const { jobState, logger } = context;
+  const { jobState } = context;
   await jobState.iterateEntities(
     { _type: bindingEntities.BINDINGS._type },
     async (bindingEntity: BindingEntity) => {
-      const cloudResourceIdentifier = traverseCloudResourcesMap(
-        bindingEntity.resource,
-      );
-      if (!cloudResourceIdentifier) {
-        logger.warn(
-          { resource: bindingEntity.resource },
-          'unable to find google cloud resource identifier.',
-        );
+      const { type, key } =
+        getTypeAndKeyFromResourceIdentifier(context, bindingEntity.resource) ??
+        {};
+      if (typeof type !== 'string' || typeof key !== 'string') {
         return;
       }
-      const targetResourceType =
-        GOOGLE_RESOURCE_TO_J1_TYPE_MAP[cloudResourceIdentifier];
-      if (!targetResourceType) {
-        logger.warn(
-          { resource: bindingEntity.resource, cloudResourceIdentifier },
-          'unable to find J1 type from google cloud resource.',
-        );
-        return;
-      } else if (targetResourceType === NONE) {
-        logger.warn(
-          { resource: bindingEntity.resource, cloudResourceIdentifier },
-          'There is no JupiterOne entity for this resource',
-        );
-        return;
-      }
-      const keys = uniq(
-        Array.isArray(targetResourceType)
-          ? targetResourceType
-          : [targetResourceType],
-      ).map((type) => getTargetKey(type, bindingEntity.resource));
-      if (keys.length < 1) {
-        logger.warn(
-          {
-            resource: bindingEntity.resource,
-            cloudResourceIdentifier,
-            targetResourceType,
-            keys,
-          },
-          'unable to generate key for type.',
-        );
-        return;
-      }
-      for (const key of keys) {
-        if (typeof key === 'string') {
-          const existingEntity = await jobState.findEntity(key);
-          if (existingEntity) {
-            await jobState.addRelationship(
-              createDirectRelationship({
-                from: bindingEntity,
-                _class: RelationshipClass.ALLOWS,
-                to: existingEntity,
-              }),
-            );
-          } else {
-            await jobState.addRelationship(
-              createMappedRelationship({
-                _class: BINDING_ALLOWS_ANY_RESOURCE_RELATIONSHIP._class,
-                _type: BINDING_ALLOWS_ANY_RESOURCE_RELATIONSHIP._type,
-                _mapping: {
-                  relationshipDirection: RelationshipDirection.FORWARD,
-                  sourceEntityKey: bindingEntity._key,
-                  targetFilterKeys: [['_type', '_key']],
-                  skipTargetCreation: false,
-                  targetEntity: {
-                    _type: Array.isArray(targetResourceType)
-                      ? targetResourceType[0]
-                      : targetResourceType, // TODO: fix
-                    _key: key,
-                  },
+      const existingEntity = await jobState.findEntity(key);
+      await jobState.addRelationship(
+        existingEntity
+          ? createDirectRelationship({
+              from: bindingEntity,
+              _class: RelationshipClass.ALLOWS,
+              to: existingEntity,
+            })
+          : createMappedRelationship({
+              _class: BINDING_ALLOWS_ANY_RESOURCE_RELATIONSHIP._class,
+              _type: BINDING_ALLOWS_ANY_RESOURCE_RELATIONSHIP._type,
+              _mapping: {
+                relationshipDirection: RelationshipDirection.FORWARD,
+                sourceEntityKey: bindingEntity._key,
+                targetFilterKeys: [['_type', '_key']],
+                skipTargetCreation: false,
+                targetEntity: {
+                  _type: type,
+                  _key: key,
                 },
-              }),
-            );
-          }
-        }
-      }
+              },
+            }),
+      );
     },
   );
 }
